move towards full version string
