+++ protocol-designer/package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "@opentrons/components": "3.7.0",
     "classnames": "^2.2.5",
+    "compare-versions": "^3.4.0",
     "formik": "^1.3.1",
     "i18next": "^11.5.0",
     "lodash": "^4.17.4",
