+++ protocol-designer/src/load-file/migration/1_1_0.js
@@ -9,11 +9,11 @@ import type {PipetteEntities} from '../../step-forms'
 import type {FormPatch} from '../../steplist/actions'
 import type {ProtocolFile, FileLabware, FilePipette} from '../../file-types'
 
-const MIGRATION_VERSION = 1
+const MIGRATION_VERSION = '1.1.0'
 
 // NOTE: these constants are copied here because
 // the default-values key did not exist for most protocols
-// pre migrationV1 in later migration files many of these values
+// pre 1.1.0 in later migration files many of these values
 // should be taken from the default-values key
 export const INITIAL_DECK_SETUP_STEP_ID: '__INITIAL_DECK_SETUP_STEP__' = '__INITIAL_DECK_SETUP_STEP__'
 export const initialDeckSetupStepForm = {
@@ -233,12 +233,12 @@ export function replaceTCDStepsWithMoveLiquidStep (fileData: ProtocolFile): Prot
   }
 }
 
-export function updateMigrationVersion (fileData: ProtocolFile): ProtocolFile {
+export function updateVersion (fileData: ProtocolFile): ProtocolFile {
   return {
     ...fileData,
     'designer-application': {
       ...fileData['designer-application'],
-      migrationVersion: MIGRATION_VERSION,
+      applicationVersion: MIGRATION_VERSION,
     },
   }
 }
@@ -248,7 +248,7 @@ const migrateFile = (fileData: ProtocolFile) => flow([
   addInitialDeckSetupStep,
   updateStepFormKeys,
   replaceTCDStepsWithMoveLiquidStep,
-  updateMigrationVersion,
+  updateVersion,
 ])(fileData)
 
 export default {
