+++ protocol-designer/src/load-file/migration/__tests__/1_1_0.test.js
@@ -11,7 +11,7 @@ import {
   MIX_DEPRECATED_FIELD_NAMES,
   replaceTCDStepsWithMoveLiquidStep,
   updateMigrationVersion,
-} from '../migrationV1.js'
+} from '../1_1_0.js'
 
 describe('renameOrderedSteps', () => {
   const migratedFile = renameOrderedSteps(oldProtocol)
